requirements.txt
asgiref==3.11.0
channels==4.3.2
channels_redis==4.3.0
Django==5.0.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.2
PyJWT==2.11.0
redis==7.1.1
sqlparse==0.5.5
tzdata==2025.3
